requirements.txt
asttokens==2.2.1
backcall==0.2.0
click==8.1.3
colorama==0.4.6
comm==0.1.2
debugpy==1.6.6
decorator==5.1.1
executing==1.2.0
Flask==2.2.2
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
ipykernel==6.16.2
ipython==7.34.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.0.0a1
jupyter_core==5.0.0rc2
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mysql-connector-python==8.0.32
mysqlclient==2.1.1
nest-asyncio==1.5.6
packaging==23.0
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.0.0
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
pure-eval==0.2.2
Pygments==2.14.0
PyMySQL==1.0.2
python-dateutil==2.8.2